add utilities to build and debug the STM32g0 flash algorithm
These scripts are specific to the STM32g0 flash algorithm.

Since the region is not memory mapped on the STM32, we need to override
some portions of the flash region definition to use a custom memory read
function (which is done in the pyocd_config.py file). The build-flm file
can be used to build the flash loader algorithm, and debug-flm can be
used to load the flm for debugging. When debug-flm is used, the FLM
should include an infinite loop to halt execution at the location the
user wishes to debug.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>