implement spi_combo flash algorithm for STM32
Implement a spi_combo flash algorithm for the STM32. This algorithm
is similar to the SPI1 flash algorithm, but performs data transfer on
the SPI2 MISO/MOSI lines. However, is is designed for use on the P300
board series, which has an issue where the SPI2 SCLK signal is not routed.
Therefore, we toggle SPI1 SCLK in sequence with SPI2 DR writes, so that the
flash chip can still be accessed.

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>